refactor: boost post button to boost button
